Move config parsing helpers and flags to platform layer
Implements sub-issue #2157 - Move config parsing infrastructure from NES module to platform layer to enable code reuse by GB and future systems.

Changes:
- Moved ParseResult enum from nes::console to platform::config
- Moved 10 parsing helper functions to platform::config:
  * parse_bool_arg, has_negation_flag, parse_bool
  * parse_u32_arg, parse_i64_arg, parse_f32_arg
  * parse_hex_u8, parse_hex_addr, parse_breakpoint_list, parse_search_paths
- Created PLATFORM_CLI_FLAGS (~39 flags for audio, video, debugging, autorun, etc.)
- Updated NES CLI_FLAGS to only contain NES-specific flags (~27 flags)
- Updated all_cli_flags() to chain PLATFORM + NES + GB flags
- Updated NES config imports to use platform helpers

Verification:
- All 8666 tests pass (no regressions)
- Help output unchanged (only flag ordering differs slightly)
- All pre-merge checkpoints pass:
  * cargo clippy --all-targets --all-features -- -D warnings ✓
  * cargo fmt ✓
  * cargo test --no-default-features --lib ✓
  * wasm-pack test --headless --chrome ✓
  * Python tests (201) ✓
  * npm test (298) ✓

This establishes the foundation for phase 2 (splitting apply_args) and phase 3 (FrontendConfig parsing).

Related to #2154
